refactor: move command.lua out of src/
