+++ lua/neotest-kotlin/init.lua
@@ -1,7 +1,7 @@
 local lib = require("neotest.lib")
 local async = require("neotest.async")
 local treesitter = require("neotest-kotlin.treesitter")
-local command = require("neotest-kotlin.src.command")
+local command = require("neotest-kotlin.command")
 local filter = require("neotest-kotlin.filter")
 local output_parser = require("neotest-kotlin.output_parser")
